+++ src/shared/Threading/LegacyThreading.h
@@ -22,11 +22,10 @@
 #define _THREADING_H
 
 // We need assertions.
-#include <Errors.h>
+#include "Debugging/Errors.h"
 
 // Platform Specific Lock Implementation
 #include "Mutex.h"
-
 
 // Platform Specific Thread Base
 #include "LegacyThreadBase.h"
