+++ src/world/Movement/MovementGenerators/SplineChainMovementGenerator.cpp
@@ -8,7 +8,7 @@ This file is released under the MIT license. See README-MIT for more information
 #include "SplineChainMovementGenerator.h"
 #include "Objects/Units/Creatures/Creature.h"
 #include "Objects/Units/Creatures/AIInterface.h"
-#include "Errors.h"
+#include "Debugging/Errors.h"
 #include "Movement/MovementManager.h"
 #include "Movement/MovementDefines.h"
 #include "Movement/Spline/MoveSpline.h"
