+++ src/world/Server/UpdateMask.h
@@ -21,7 +21,7 @@
 #ifndef UPDATEMASK_H
 #define UPDATEMASK_H
 
-#include "Errors.h"
+#include "Debugging/Errors.h"
 #include <cstring>
 #include "CommonTypes.hpp"
