+++ src/world/Server/WorldRunnable.cpp
@@ -3,10 +3,8 @@ Copyright (c) 2014-2023 AscEmu Team <http://www.ascemu.org>
 This file is released under the MIT license. See README-MIT for more information.
 */
 
-
-
+#include "Debugging/CrashHandler.h"
 #include "WorldRunnable.h"
-#include <CrashHandler.h>
 #include "World.h"
 #include "ServerState.h"
